requirements.txt
opencv-python
mediapipe
pyserial
PyQt6
